Drop pid files and fix working dirs
When using systemd, the pid files are actually not important so they can
be dropped.

The the WorkingDirectory is set to ~ which will resolve to the user's
home directory. That means systemd will ensure it is mounted before
starting services. Pulp itself doesn't use it so it's a cheap way for
a more reliable service start.

Subscribing to pulpcore-worker@.service is another bugfix to ensure all
workers use the correct service definition.

Explicitly setting the type and group is more for consistency and
doesn't actually change anything.