Make REMOTE_USER_ENVIRON_NAME mandatory
Pulpcore has a default for this (which is copied). That means it's a
good idea to always set it. Otherwise the illusion is created that it's
an optional thing. However, the reverse proxy must be set up to use
this. If the header is not cleared, it's a security risk because an
attack can spoof the header.