Make it mostly compatible with gha-puppet
In this repository multiple test suites run. That is currently not
supported by gha-puppet so the file becomes unmanaged. It does take the
improvement of a test at the end, which allows automerge to be enabled.
It also keeps the Gemfile in sync and aligns on PUPPET_GEM_VERSION.

At the same time it increases the cache version to work around the
broken Ruby 2.5 cache which includes Psych 4.